+++ packages/web3-modal/src/components/modals/ConnectionModal/ConnectedCheckMark.tsx
@@ -9,7 +9,7 @@ export const StyledConnectedCheckMark = styled(RowStart)`
 
   gap: 0.4em;
 
-  font-size: 0.5em;
+  font-size: 0.4em;
 `
 
 export function ConnectedCheckMark() {